Remove and consolidate package
