+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/profiler/JfrProfiler.java
@@ -1,12 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.applicationinsights.agent.internal.profiler.profiler;
+package com.microsoft.applicationinsights.agent.internal.profiler;
 
 import com.microsoft.applicationinsights.agent.internal.configuration.Configuration;
-import com.microsoft.applicationinsights.agent.internal.profiler.ProfileHandler;
-import com.microsoft.applicationinsights.agent.internal.profiler.ProfilerConfiguration;
-import com.microsoft.applicationinsights.agent.internal.profiler.ProfilerConfigurationHandler;
 import com.microsoft.applicationinsights.agent.internal.profiler.upload.UploadCompleteHandler;
 import com.microsoft.applicationinsights.alerting.alert.AlertBreach;
 import com.microsoft.applicationinsights.alerting.config.AlertConfiguration;
