+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/profiler/ProfileTypes.java
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.microsoft.applicationinsights.agent.internal.profiler.profiler;
+package com.microsoft.applicationinsights.agent.internal.profiler;
 
 enum ProfileTypes {
   PROFILE,
